improve snapshot loading script

+++ bin/load_snapshot.sh
@@ -6,13 +6,14 @@ set -euo pipefail
 DIR=$( cd -- "$( dirname -- "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )
 SNAPSHOT=${1:-}  # if no snapshot path is provided, it will load directly from POSTGRES_SNAPSHOT_SRC_URL
 
-[ -f "$DIR/../.env" ] && source "$DIR/../.env"
+# only load env vars from .env file if POSTGRES_URL is not set
+[ -z "${POSTGRES_URL:-}" ] && [ -f "$DIR/../.env" ] && source "$DIR/../.env"
 
 # to load into prod, set POSTGRES_URL to the prod db url and pass a snapshot file as an argument
 # to load into dev, set POSTGRES_URL to the dev db url and POSTGRES_SNAPSHOT_SRC_URL to the prod db url (or pass a snapshot file as an argument)
 
 [ -z "${POSTGRES_URL:-}" ] && echo "\$POSTGRES_URL must be set" && exit 1
-[ -z "${POSTGRES_SNAPSHOT_SRC_URL:-}" ] && echo "\$POSTGRES_SNAPSHOT_SRC_URL must be set" && exit 1
+[ -z "${SNAPSHOT}" ] && [ -z "${POSTGRES_SNAPSHOT_SRC_URL:-}" ] && echo "\$POSTGRES_SNAPSHOT_SRC_URL must be set or a snapshot file must be provided as an argument" && exit 1
 
 if [[ $POSTGRES_URL =~ postgres(ql)?://([A-Za-z0-9]+):([^@]+)@([A-Za-z0-9.-]+)(:([0-9]+))?/([A-Za-z0-9]+) ]]; then
     user="${BASH_REMATCH[2]}"
@@ -33,21 +34,25 @@ connections=$(PGPASSWORD="$pass" psql -U "$user" -h "$host" -p "$port" postgres 
 [ "$connections" -gt 1 ] && echo "Close all other active db connections before loading snapshot" && exit 1
 
 
-echo "testing remote connection before dropping local db"
-pg_dump "$POSTGRES_SNAPSHOT_SRC_URL" --schema-only -t pg_namespace > /dev/null || (echo "Failed to connect to snapshot source db" && exit 1)
+if [ -z "$SNAPSHOT" ]; then
+  echo "testing remote connection before dropping local db"
+  pg_dump "$POSTGRES_SNAPSHOT_SRC_URL" --schema-only -t pg_namespace > /dev/null || (echo "Failed to connect to snapshot source db" && exit 1)
+fi
 
 
+echo "dropping and recreating db"
 PGPASSWORD="$pass" psql -U "$user" -h "$host" -p "$port" postgres \
   -c "DROP DATABASE IF EXISTS $db;" \
   -c "CREATE DATABASE $db;"
 
 
-echo "loading snapshot"
 postgres_url_trimmed="${POSTGRES_URL%%\?*}" ## Remove options query string
 if [ -n "$SNAPSHOT" ]; then
-    cat "$SNAPSHOT" | psql "$postgres_url_trimmed"
+  echo "loading snapshot from ${SNAPSHOT}"
+  cat "$SNAPSHOT" | psql "$postgres_url_trimmed"
 else
-    pg_dump "$POSTGRES_SNAPSHOT_SRC_URL" | psql "$postgres_url_trimmed"
+  echo "loading snapshot from src db"
+  pg_dump "$POSTGRES_SNAPSHOT_SRC_URL" | psql "$postgres_url_trimmed"
 fi
